Renamed the example tuner plugin to model_demo to accommodate the nccl example tuner plugin.
